refact: Refatoração e reorganização do escopo de bookmarket
