+++ src/main/java/servico/bookmarket/exceptions/UmbrellaException.java
@@ -1,0 +1,14 @@
+package servico.bookmarket.exceptions;
+
+public class UmbrellaException extends RuntimeException {
+
+    /**
+     * Default Generated SerialUID
+     */
+    private final long serialVersionUID = 8525983676993371110L;
+
+    public UmbrellaException(Exception e) {
+        super(e);
+    }
+
+}
