+++ src/main/java/servico/bookmarket/interfaces/Action.java
@@ -1,0 +1,6 @@
+package servico.bookmarket.interfaces;
+
+public interface Action<STATE> {
+
+    Object executeOn(STATE sm);
+}
